drivers/netxml-ups.c: fix casting ssize_t value up to size_t for ne_buffer_append() args

+++ drivers/netxml-ups.c
@@ -1650,7 +1650,7 @@ static int send_http_request(
 				break;
 
 			if (NULL != resp_body)
-				ne_buffer_append(resp_body, buff, read);
+				ne_buffer_append(resp_body, buff, (size_t)read);
 		}
 
 		if (NE_OK != status) {
